+++ ead/ead_test.go
@@ -15,6 +15,7 @@ var testFixturePath string = filepath.Join(".", "testdata")
 var omegaTestFixturePath string = filepath.Join(testFixturePath, "omega", "v0.1.5")
 var falesTestFixturePath string = filepath.Join(testFixturePath, "fales")
 var nyhsTestFixturePath string = filepath.Join(testFixturePath, "nyhs")
+var akkasahTestFixturePath string = filepath.Join(testFixturePath, "akkasah")
 
 func testProcessDID(did *DID) {
 	daos := did.DAO
@@ -83,7 +84,7 @@ func getOmegaEAD(t *testing.T) EAD {
 }
 
 func getFalesMSS460EAD(t *testing.T) EAD {
-	EADXML, err := ioutil.ReadFile(falesTestFixturePath + "/" + "/mss_460.xml")
+	EADXML, err := ioutil.ReadFile(falesTestFixturePath + "/" + "mss_460.xml")
 	failOnError(t, err, "Unexpected error")
 
 	var ead EAD
@@ -95,6 +96,17 @@ func getFalesMSS460EAD(t *testing.T) EAD {
 
 func getNYHSFoundling(t *testing.T) EAD {
 	EADXML, err := ioutil.ReadFile(nyhsTestFixturePath + "/" + "nyhs_foundling.xml")
+	failOnError(t, err, "Unexpected error")
+
+	var ead EAD
+	err = xml.Unmarshal([]byte(EADXML), &ead)
+	failOnError(t, err, "Unexpected error")
+
+	return ead
+}
+
+func getAkkasahADMC030REF184EAD(t *testing.T) EAD {
+	EADXML, err := ioutil.ReadFile(akkasahTestFixturePath + "/" + "ad_mc_030_ref184.xml")
 	failOnError(t, err, "Unexpected error")
 
 	var ead EAD
@@ -502,7 +514,7 @@ func TestFilteredStringStringFunction(t *testing.T) {
 
 func TestGetConvertedTextWithTags(t *testing.T) {
 	t.Run("Test GetConvertedTextWithTags()", func(t *testing.T) {
-		sut, _ := GetConvertedTextWithTags(`The Young Devil 
+		sut, _ := GetConvertedTextWithTags(`The Young Devil
 | Actors in Image: Ahmad Ramzy, Amal Farid, Hussein Riad, Zinat Sedki, Ragaa Youssef: 1958-`)
 
 		assertEqual(t, "The Young Devil | Actors in Image: Ahmad Ramzy, Amal Farid, Hussein Riad, Zinat Sedki, Ragaa Youssef: 1958-", string(sut), "Test TestGetConvertedTextWithTags()")
@@ -517,5 +529,62 @@ New York University Archives<lb/> Elmer Holmes Bobst Library<lb/> 70 Washington 
 	t.Run("Test GetConvertedTextWithTagsNoLBConversion()", func(t *testing.T) {
 		sut, _ := GetConvertedTextWithTagsNoLBConversion(input)
 		assertEqual(t, want, string(sut), "Test TestGetConvertedTextWithTagsNoLBConversion()")
+	})
+}
+
+
+func TestJSONMarshalingInitPresentationContainersNOOP(t *testing.T) {
+	t.Run("JSON Marshaling with call to InitPresentationContainers() NOOP", func(t *testing.T) {
+		ead := getOmegaEAD(t)
+
+		ead.InitPresentationContainers()
+
+		jsonData, err := json.MarshalIndent(ead, "", "    ")
+		failOnError(t, err, "Unexpected error marshaling JSON")
+
+		// reference file includes newline at end of file so
+		// add newline to jsonData
+		jsonData = append(jsonData, '\n')
+
+		referenceFile := omegaTestFixturePath + "/" + "mos_2021.json"
+		referenceFileContents, err := ioutil.ReadFile(referenceFile)
+		failOnError(t, err, "Unexpected error reading reference file")
+
+		if !bytes.Equal(referenceFileContents, jsonData) {
+			jsonFile := "./testdata/tmp/failing-marshal-with-presentation-containers-noop.json"
+			err = ioutil.WriteFile(jsonFile, []byte(jsonData), 0644)
+			failOnError(t, err, fmt.Sprintf("Unexpected error writing %s", jsonFile))
+
+			errMsg := fmt.Sprintf("JSON Data does not match reference file.\ndiff %s %s", jsonFile, referenceFile)
+			t.Errorf(errMsg)
+		}
+	})
+}
+
+func TestJSONMarshalingInitPresentationContainers(t *testing.T) {
+	t.Run("JSON Marshaling with call to InitPresentationContainers()", func(t *testing.T) {
+		ead := getAkkasahADMC030REF184EAD(t)
+
+		ead.InitPresentationContainers()
+
+		jsonData, err := json.MarshalIndent(ead, "", "    ")
+		failOnError(t, err, "Unexpected error marshaling JSON")
+
+		// reference file includes newline at end of file so
+		// add newline to jsonData
+		jsonData = append(jsonData, '\n')
+
+		referenceFile := akkasahTestFixturePath + "/" + "ad_mc_030_ref184.json"
+		referenceFileContents, err := ioutil.ReadFile(referenceFile)
+		failOnError(t, err, "Unexpected error reading reference file")
+
+		if !bytes.Equal(referenceFileContents, jsonData) {
+			jsonFile := "./testdata/tmp/failing-marshal-with-presentation-containers.json"
+			err = ioutil.WriteFile(jsonFile, []byte(jsonData), 0644)
+			failOnError(t, err, fmt.Sprintf("Unexpected error writing %s", jsonFile))
+
+			errMsg := fmt.Sprintf("JSON Data does not match reference file.\ndiff %s %s", jsonFile, referenceFile)
+			t.Errorf(errMsg)
+		}
 	})
 }