+++ ead/ead.go
@@ -5,7 +5,7 @@ package ead
 // Based on: "Data model for parsing EAD <archdesc> elements": https://jira.nyu.edu/jira/browse/FADESIGN-29.
 
 const (
-	Version = "0.15.2-stream-parsing-2023-01-16T19_55"
+	Version = "0.15.2-stream-parsing-2023-01-17T17_50"
 )
 
 type EAD struct {